requirements.txt
amqp==1.4.9
anyjson==0.3.3
apturl==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.3.0.23
blinker==1.4
Brlapi==0.7.0
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.3
django-allauth==0.51.0
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-celery==3.3.1
django-ckeditor==6.5.1
django-comments-xtd==2.9.9
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-dbbackup==4.0.2
django-decorator-include==3.0
django-etc==1.4.0
django-froala-editor==4.0.16
django-hitcount==1.3.5
django-js-asset==2.0.0
django-newsletter-subscription==0.5.0
django-pandas==0.6.6
django-paypal==2.0
django-social-share==2.3.0
django-storages==1.13.1
django-taggit==3.1.0
django-tinymce==3.5.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
docutils==0.19
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kombu==3.0.37
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
markdown2==2.4.6
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpy==1.24.1
oauthlib==3.1.0
olefile==0.46
pandas==1.5.2
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.4
protobuf==3.6.1
psycopg2==2.9.5
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyquery==1.4.3
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.7
pyxdg==0.26
PyYAML==5.3.1
readtime==2.0.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.12
sqlalchemy-migrate==0.13.0
sqlparse==0.2.4
systemd-python==234
Tempita==0.5.2
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.17.0
wadllib==1.3.3
xkit==0.0.0
